Fixes #1 by checking for fireplace connection.
This should fix #1 by checking for an active fireplace connection before
trying to use it. When there is no fireplace connection data cannot be
loaded from Clojars on plugin load so data loading is deferred until the
first request is made to hot load a dependency. On plugin load we also
try to "figure out" if there is a dev namespace, since we can't do this
without a fireplace connection we default to the user namespace.